+++ agentid/__init__.py
@@ -1,3 +1,3 @@
 """AgentAction package."""
 
-__version__ = "0.29.0rc1"
+__version__ = "0.29.0rc2"